% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-ampute-data.R
\name{AmputeData}
\alias{AmputeData}
\title{Ampute Data}
\usage{
AmputeData(data_complete, mech = "MAR", prop = 0.3, patterns = NULL)
}
\arguments{
\item{data_complete}{Numeric matrix.
Output of the \code{GenData} function
or a three-column data set with complete data.}

\item{mech}{Missing data mechanism.}

\item{prop}{Proportion of missing data.}

\item{patterns}{Numeric matrix consisting of zeroes and ones.
Each row in the matrix represents a missing data pattern
where 0 indicates a missing observation.
If \code{patterns = NULL},
the default value is all possible missing data patterns
for a data set with 3 columns.}
}
\value{
Returns a dataframe.
}
\description{
The function generates a data set with missing values
for data generated by \code{GenData}
using the multivariate amputation approach.
See \code{mice::ampute()} for more details.
}
\references{
Schouten, R. M., Lugtig, P. and Vink, G. (2018).
Generating missing values for simulation purposes:
A multivariate amputation procedure.
\emph{Journal of Statistical Computation and Simulation},
\emph{88}(15), 1909--1930.
\doi{10.1080/00949655.2018.1491577}
}
\seealso{
Other Data Generation Functions: 
\code{\link{GenData}()},
\code{\link{ImputeData}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Data Generation Functions}
\keyword{ampute}
\keyword{gendata}
\keyword{manMCMedMiss}
\keyword{mediation}
