% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-check.R
\name{Check}
\alias{Check}
\title{Check Replication}
\usage{
Check(taskids, repids, output_folder, nbmi)
}
\arguments{
\item{taskids}{Vector of taskids.}

\item{repids}{Vector of repids.}

\item{output_folder}{Character string.
Output folder.}

\item{nbmi}{Logical.
If \code{nbmi = TRUE}, include \code{MINB}, \code{NBMI}, and \code{StackedMINB}.}
}
\value{
Returns file names of replications that need to be run.
}
\description{
Check Replication
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{check}
\keyword{manMCMedMiss}
