% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-check-gen-data.R
\name{CheckGenData}
\alias{CheckGenData}
\title{Check Replication - SimGenData}
\usage{
CheckGenData(taskid, repid, output_folder, suffix)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{repid}{Positive integer.
Replication ID.}

\item{output_folder}{Character string.
Output folder.}

\item{suffix}{Character string.
Output of \code{manMCMedMiss:::.SimSuffix()}.}
}
\value{
Returns file names of replications that need to be run.
}
\description{
Check Replication - SimGenData
}
\details{
This function is executed via the \code{Check} function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{check}
\keyword{gendata}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{normal}
