% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-check-impute-data.R
\name{CheckImputeData}
\alias{CheckImputeData}
\title{Check Replication - SimImputeData}
\usage{
CheckImputeData(taskid, repid, output_folder, suffix)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{repid}{Positive integer.
Replication ID.}

\item{output_folder}{Character string.
Output folder.}

\item{suffix}{Character string.
Output of \code{manMCMedMiss:::.SimSuffix()}.}
}
\value{
Returns file names of replications that need to be run.
}
\description{
Check Replication - SimImputeData
}
\details{
This function is executed via the \code{Check} function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{check}
\keyword{gendata}
\keyword{impute}
\keyword{manMCMedMiss}
