% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-check-joint-sig-ml.R
\name{CheckJointSigML}
\alias{CheckJointSigML}
\title{Check Replication - SimJointSigML}
\usage{
CheckJointSigML(taskid, repid, output_folder, suffix)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{repid}{Positive integer.
Replication ID.}

\item{output_folder}{Character string.
Output folder.}

\item{suffix}{Character string.
Output of \code{manMCMedMiss:::.SimSuffix()}.}
}
\value{
Returns file names of replications that need to be run.
}
\description{
Check Replication - SimJointSigML
}
\details{
This function is executed via the \code{Check} function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{check}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{ml}
\keyword{test}
