% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-check-nb-mi.R
\name{CheckNBMI}
\alias{CheckNBMI}
\title{Check Replication - SimNBMI}
\usage{
CheckNBMI(taskid, repid, output_folder, suffix)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{repid}{Positive integer.
Replication ID.}

\item{output_folder}{Character string.
Output folder.}

\item{suffix}{Character string.
Output of \code{manMCMedMiss:::.SimSuffix()}.}
}
\value{
Returns file names of replications that need to be run.
}
\description{
Check Replication - SimNBMI
}
\details{
This function is executed via the \code{Check} function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{check}
\keyword{ci}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{mi}
