% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-elliot-2007.R
\docType{data}
\name{elliot2007}
\alias{elliot2007}
\title{Intervention on Healthy Dietary Behavior
Via Knowledge of Healthy Dietary Behavior}
\format{
A data frame with 354 cases and 3 variables:
\describe{
\item{x}{
Intervention group membership.
}
\item{m}{
Knowledge of healthy dietary behavior
(knowledge post-intervention minus knowledge pre-intervention).
}
\item{y}{
Healthy dietary behavior
(behavior post-intervention minus behavior pre-intervention).
}
}
}
\source{
\url{https://odin.mdacc.tmc.edu/~yyuan/Software/Mediation/Bayesian/CodeRelease/BayesianMediation/firefighters.txt} # nolint
}
\usage{
elliot2007
}
\description{
A study by Elliot and others (2007)
on the effects of an intervention on healthy dietary behavior
mediated by knowledge of healthy dietary behavior.
This data was used as an empirical example in Yuan and MacKinnon (2009) and
Wu and Jia (2013).
}
\references{
Elliot, D. L., Goldberg, L., Kuehl, K. S., Moe, E. L., Breger, R. K., &
Pickering, M. A. (2007).
The PHLAME (Promoting Healthy Lifestyles: Alternative Models' Effects)
firefighter study: Outcomes of two models of behavior change.
\emph{Journal of occupational and environmental medicine}, \emph{49}(2), 204–213.
\doi{10.1097/JOM.0b013e3180329a8d}

Wu, W., & Jia, F. (2013).
A new procedure to test mediation with missing data
through nonparametric bootstrapping and multiple imputation,
\emph{Multivariate Behavioral Research}, \emph{48}(5), 663-691.
\doi{10.1080/00273171.2013.816235}

Yuan, Y., & MacKinnon, D. P. (2009).
Bayesian mediation analysis.
\emph{Psychological methods}, \emph{14}(4), 301–322.
\doi{10.1037/a0016972}
}
\keyword{data}
