% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-fit-model-mi.R
\name{FitModelMI}
\alias{FitModelMI}
\title{Fit the Simple Mediation Model using Multiple Imputation}
\usage{
FitModelMI(data_mi, mplus_bin)
}
\arguments{
\item{data_mi}{List of numeric matrices.
Output of the \code{ImputeData} function
or a list of three-column data sets with imputed data.}

\item{mplus_bin}{Character string.
Path of Mplus binary.}
}
\value{
Returns a list with the following elements:
\describe{
\item{coef}{
Vector of pooled coefficients/parameter estimates
\eqn{\bar{\boldsymbol{\theta}}}.
}
\item{vcov}{
Total covariance matrix
\eqn{\mathbf{V}_{\mathrm{total}}}.
}
\item{vcov_tilde}{
Adjusted total covariance matrix
\eqn{\tilde{\mathbf{V}}_{\mathrm{total}}}.
}
\item{vcov_between}{
Covariance between imputations
\eqn{\mathbf{V}_{\mathrm{between}}}.
}
\item{vcov_within}{
Covariance within imputations
\eqn{\mathbf{V}_{\mathrm{within}}}.
}
\item{ariv}{
Average relative increase in variance
\eqn{\mathrm{ARIV}}.
}
\item{m}{
Number of imputations
\eqn{M}.
}
\item{k}{
Number of parameters
\eqn{k}.
}
\item{nu1}{
Numerator degrees of freedom
\eqn{\nu_1} for \eqn{D_1}.
}
\item{nu2}{
Denominator degrees of freedom
\eqn{\nu_2} for \eqn{D_1}.
}
\item{d1}{
\eqn{D_1} test statistic.
}
}
}
\description{
Fit the Simple Mediation Model using Multiple Imputation
}
\details{
Let
\deqn{M = \textrm{Number of imputations}, \quad \textrm{and}}
\deqn{m = \left\{ 1, 2, \cdots, M \right\}.}
The vector of pooled coefficients/parameter estimates is given by
\deqn{
    \bar{\boldsymbol{\theta}}
    =
    M^{-1}
    \sum_{m = 1}^{M}
    \hat{\boldsymbol{\theta}}_{m} .
}
The covariance within imputations is given by
\deqn{
    \mathbf{V}_{\mathrm{within}}
    =
    M^{-1}
    \sum_{m = 1}^{M}
    \mathrm{Var}
    \left(
    \hat{\boldsymbol{\theta}}_{m}
    \right)
}
where \eqn{\mathrm{Var} \left( \hat{\boldsymbol{\theta}}_{m} \right)}
is the parameter covariance matrix for the \eqn{m^{\mathrm{th}}} imputation.
The covariance between imputations is given by
\deqn{
    \mathbf{V}_{\mathrm{between}}
    =
    \left(
    M - 1
    \right)^{-1}
    \sum_{m = 1}^{M}
    \left(
    \hat{\boldsymbol{\theta}}_{m}
    -
    \bar{\boldsymbol{\theta}}
    \right)
    \left(
    \hat{\boldsymbol{\theta}}_{m}
    -
    \bar{\boldsymbol{\theta}}
    \right)^{\prime} .
}
The total covariance matrix is given by
\deqn{
    \mathbf{V}_{\mathrm{total}}
    =
    \mathbf{V}_{\mathrm{within}}
    +
    \mathbf{V}_{\mathrm{between}}
    +
    M^{-1}
    \mathbf{V}_{\mathrm{between}} .
}
The adjusted total covariance matrix is given by
\deqn{
    \tilde{\mathbf{V}}_{\mathrm{total}}
    =
    \left(
    1
    +
    \mathrm{ARIV}
    \right)
    \mathbf{V}_{\mathrm{within}}
}
where
\deqn{
    \mathrm{ARIV}
    =
    k^{-1}
    \left[
    \left(
    1
    +
    M^{-1}
    \right)
    \mathrm{tr}
    \left(
    \mathbf{V}_{\mathrm{between}}
    \mathbf{V}_{\mathrm{within}}^{-1}
    \right)
    \right]
}
and \eqn{k} is the number of parameters.
}
\references{
Li, K. H., Raghunathan, T. E., & Rubin, D. B. (1991).
Large-sample significance levels from multiply imputed data
using moment-based statistics and an F reference distribution.
\emph{Journal of the American Statistical Association}, 86 (416), 1065–1073.
\doi{10.1080/01621459.1991.10475152}

Rubin, D. B. (1987).
\emph{Multiple imputation for nonresponse in surveys}.
John Wiley & Sons, Inc.
\doi{10.1002/9780470316696}
}
\seealso{
Other Model Fitting Functions: 
\code{\link{FitModelIndirect}()},
\code{\link{FitModelML}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Model Fitting Functions}
\keyword{fit}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{mi}
