% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-fit-model-ml.R
\name{FitModelML}
\alias{FitModelML}
\title{Fit the Simple Mediation Model using Normal Theory Maximum Likelihood}
\usage{
FitModelML(data, mplus_bin)
}
\arguments{
\item{data}{Numeric matrix.
Output of the \code{GenData} or \code{AmputeData} functions.}

\item{mplus_bin}{Character string.
Path of Mplus binary.}
}
\value{
Returns a list with the following elements:
\describe{
\item{fit}{
Model fit.
}
\item{coef}{
Coefficients/parameter estimates
}
\item{vcov}{
Sampling variance-covariance matrix.
}
\item{output}{
Mplus output.
}
}
Values in \code{fit}.
\describe{
\item{free_parameters}{
Number of free parameters
}
\item{h0_loglikelihood}{
H0 loglikelihood.
}
\item{h1_loglikelihood}{
H1 loglikelihood.
}
\item{aic}{
Akaike information criterion (AIC).
}
\item{bic}{
Bayesian information criterion (BIC).
}
\item{sabic}{
Sample-size adjusted BIC (SABIC).
}
\item{chisq}{
Chi-square value.
}
\item{chisq_df}{
Chi-square degrees of freedom.
}
\item{chisq_p}{
Chi-square p-value.
}
\item{cfi}{
Comparative fit index (CFI).
}
\item{tli}{
Tucker–Lewis index (TLI).
}
\item{rmsea}{
Root mean square error of approximation (RMSEA)
estimate.
}
\item{rmsea_low}{
Root mean square error of approximation (RMSEA)
lower limit confidence interval.
}
\item{rmsea_up}{
Root mean square error of approximation (RMSEA)
upper limit confidence interval.
}
\item{rmsea_p}{
Root mean square error of approximation (RMSEA)
probability.
}
\item{srmr}{
Standardized root mean square residual (SRMR).
}
\item{condition_number}{
Condition number for the information matrix
(ratio of smallest to largest eigenvalue).
}
}
}
\description{
Fit the Simple Mediation Model using Normal Theory Maximum Likelihood
}
\seealso{
Other Model Fitting Functions: 
\code{\link{FitModelIndirect}()},
\code{\link{FitModelMI}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Model Fitting Functions}
\keyword{fit}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{ml}
