% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-impute-data.R
\name{ImputeData}
\alias{ImputeData}
\title{Impute Data}
\usage{
ImputeData(data_missing, m, mplus_bin)
}
\arguments{
\item{data_missing}{Numeric matrix.
Output of the \code{AmputeData} function
or a three-column data set with missing data.}

\item{m}{Positive integer.
Number of imputations.}

\item{mplus_bin}{Character string.
Path of Mplus binary.}
}
\value{
Returns a list of complete data sets.
}
\description{
Generates \code{m} complete data sets using multiple imputation.
}
\references{
Asparouhov, T., & Muthen, B. (2022).
\emph{Multiple imputation with Mplus}.
Retrieved from \url{http://www.statmodel.com/download/Imputations7.pdf}
}
\seealso{
Other Data Generation Functions: 
\code{\link{AmputeData}()},
\code{\link{GenData}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Data Generation Functions}
\keyword{gendata}
\keyword{impute}
\keyword{manMCMedMiss}
