% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-joint-sig-mi.R
\name{JointSigMI}
\alias{JointSigMI}
\title{Joint Significance Test for the Indirect Effect using Multiple Imputation}
\usage{
JointSigMI(fit_mi, alpha = c(0.05, 0.01, 0.001))
}
\arguments{
\item{fit_mi}{Object.
Output of the \code{FitModelMI} function.}

\item{alpha}{Numeric vector.
Significance level.}
}
\value{
Returns a list with the following elements:
\describe{
\item{vcov}{
Joint significance test based on the total covariance matrix.
}
\item{vcov_tilde}{
Joint significance test based on the adjusted total covariance matrix.
}
}
Note the \code{1} corresponds to a significant result and \code{0} otherwise.
The output vector's name corresponds to the \code{alpha} level.
}
\description{
Joint Significance Test for the Indirect Effect using Multiple Imputation
}
\seealso{
Other Significance Test Functions: 
\code{\link{JointSigML}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Significance Test Functions}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{mi}
\keyword{test}
