% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-joint-sig-ml.R
\name{JointSigML}
\alias{JointSigML}
\title{Joint Significance Test for the Indirect Effect
using Normal Theory Maximum Likelihood}
\usage{
JointSigML(fit_ml, alpha = c(0.05, 0.01, 0.001))
}
\arguments{
\item{fit_ml}{Object.
Output of the \code{FitModelML} function.}

\item{alpha}{Numeric vector.
Significance level.}
}
\value{
Note the \code{1} corresponds to a significant result and \code{0} otherwise.
The output vector's name corresponds to the \code{alpha} level.
}
\description{
Joint Significance Test for the Indirect Effect
using Normal Theory Maximum Likelihood
}
\seealso{
Other Significance Test Functions: 
\code{\link{JointSigMI}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Significance Test Functions}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{ml}
\keyword{test}
