% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-mc-mi.R
\name{MCMI}
\alias{MCMI}
\title{Monte Carlo Confidence Intervals for the Indirect Effect
using Multiple Imputation}
\usage{
MCMI(fit_mi, R = 20000L, alpha = c(0.05, 0.01, 0.001))
}
\arguments{
\item{fit_mi}{Object.
Output of the \code{FitModelMI} function.}

\item{R}{Positive integer.
Number of Monte Carlo replications.}

\item{alpha}{Numeric vector.
Significance level.}
}
\value{
Returns a list with the following elements:
\describe{
\item{vcov}{
Monte Carlo confidence intervals based on the total covariance matrix.
}
\item{vcov_tilde}{
Monte Carlo confidence intervals
based on the adjusted total covariance matrix.
}
}
}
\description{
Monte Carlo Confidence Intervals for the Indirect Effect
using Multiple Imputation
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{MCML}()},
\code{\link{MINB}()},
\code{\link{NBMI}()},
\code{\link{NBML}()},
\code{\link{NBStackedMI}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{mi}
