% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-mc-ml.R
\name{MCML}
\alias{MCML}
\title{Monte Carlo Confidence Intervals for the Indirect Effect
using Normal Theory Maximum Likelihood}
\usage{
MCML(fit_ml, R = 20000L, alpha = c(0.05, 0.01, 0.001))
}
\arguments{
\item{fit_ml}{Object.
Output of the \code{FitModelML} function.}

\item{R}{Positive integer.
Number of Monte Carlo replications.}

\item{alpha}{Numeric vector.
Significance level.}
}
\value{
Monte Carlo confidence intervals.
}
\description{
Monte Carlo Confidence Intervals for the Indirect Effect
using Normal Theory Maximum Likelihood
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{MCMI}()},
\code{\link{MINB}()},
\code{\link{NBMI}()},
\code{\link{NBML}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{ml}
