% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-mi-combine.R
\name{MICombine}
\alias{MICombine}
\title{Combine Parameter Estimates Vector and Sampling Variance-Covariance Matrix
Estimated from Multiple Imputations}
\usage{
MICombine(param_vec, var_mat)
}
\arguments{
\item{param_vec}{List of vectors of paramater estimates.}

\item{var_mat}{List of matrices of sampling variances and covariances.}
}
\value{
Returns a list with the following elements:
\describe{
\item{m}{
Number of imputations
\eqn{M}.
}
\item{k}{
Number of parameters
\eqn{k}.
}
\item{estimates}{
Vector of pooled coefficients/parameter estimates
\eqn{\bar{\boldsymbol{\theta}}}.
}
\item{within}{
Covariance within imputations
\eqn{\mathbf{V}_{\mathrm{within}}}.
}
\item{between}{
Covariance between imputations
\eqn{\mathbf{V}_{\mathrm{between}}}.
}
\item{total}{
Total covariance matrix
\eqn{\mathbf{V}_{\mathrm{total}}}.
}
\item{total_adjusted}{
Adjusted total covariance matrix
\eqn{\tilde{\mathbf{V}}_{\mathrm{total}}}.
}
\item{ariv}{
Average relative increase in variance
\eqn{\mathrm{ARIV}}.
}
}
}
\description{
Combine Parameter Estimates Vector and Sampling Variance-Covariance Matrix
Estimated from Multiple Imputations
}
\references{
Li, K. H., Raghunathan, T. E., & Rubin, D. B. (1991).
Large-sample significance levels from multiply imputed data
using moment-based statistics and an F reference distribution.
\emph{Journal of the American Statistical Association}, 86 (416), 1065–1073.
\doi{10.1080/01621459.1991.10475152}

Rubin, D. B. (1987).
\emph{Multiple imputation for nonresponse in surveys}.
John Wiley & Sons, Inc.
\doi{10.1002/9780470316696}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Multiple Imputation Functions}
\keyword{manMCMedMiss}
\keyword{mi}
