% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-mi-nb.R
\name{MINB}
\alias{MINB}
\title{Nonparametric Bootstrap Confidence Intervals for the Indirect Effect
using Multiple Imputation
(NB nested within MI or MI(NB))}
\usage{
MINB(data_mi, B = 5000L, alpha = c(0.05, 0.01, 0.001))
}
\arguments{
\item{data_mi}{List of numeric matrices.
Output of the \code{ImputeData} function
or a list of three-column data sets with imputed data.}

\item{B}{Positive integer.
Number of bootstrap samples.}

\item{alpha}{Numeric vector.
Significance level.}
}
\value{
Nonparametric bootstrap confidence intervals.
\code{bc} corresponds to bias-corrected and \code{pc} corresponds to percentile
confidence intervals.
}
\description{
Nonparametric Bootstrap Confidence Intervals for the Indirect Effect
using Multiple Imputation
(NB nested within MI or MI(NB))
}
\references{
Wu, Wei., & Jia, F. (2012).
A new procedure to test mediation with missing data
through nonparametric bootstrapping and multiple imputation.
\emph{Multivariate Behavioral Research}, \emph{48}(5), 663–691.
\doi{10.1080/00273171.2013.816235}
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{MCMI}()},
\code{\link{MCML}()},
\code{\link{NBMI}()},
\code{\link{NBML}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{mi}
