% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-nb-mi.R
\name{NBMI}
\alias{NBMI}
\title{Nonparametric Bootstrap Confidence Intervals for the Indirect Effect
using Multiple Imputation
(MI nested within NB or NB(MI))}
\usage{
NBMI(
  data_missing,
  data_mi,
  B = 5000L,
  m = 100L,
  alpha = c(0.05, 0.01, 0.001),
  mplus_bin
)
}
\arguments{
\item{data_missing}{Numeric matrix.
Output of the \code{AmputeData} function
or a three-column data set with missing data.}

\item{data_mi}{List of numeric matrices.
Output of the \code{ImputeData} function
or a list of three-column data sets with imputed data.}

\item{B}{Positive integer.
Number of bootstrap samples.}

\item{m}{Positive integer.
Number of imputations.}

\item{alpha}{Numeric vector.
Significance level.}

\item{mplus_bin}{Character string.
Path of Mplus binary.}
}
\value{
Nonparametric bootstrap percentile confidence intervals.
}
\description{
Nonparametric Bootstrap Confidence Intervals for the Indirect Effect
using Multiple Imputation
(MI nested within NB or NB(MI))
}
\references{
Zhang, Z., & Wang, L. (2012).
Methods for mediation analysis with missing data.
\emph{Psychometrika}, \emph{78}(1), 154–184.
\doi{10.1007/s11336-012-9301-5}

Zhang, Z., Wang, L., & Tong, X. (2015).
Mediation analysis with missing data
through multiple imputation and bootstrap.
\emph{Quantitative psychology research} (pp. 341–355).
Springer International Publishing.
\doi{10.1007/978-3-319-19977-1_24}
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{MCMI}()},
\code{\link{MCML}()},
\code{\link{MINB}()},
\code{\link{NBML}()},
\code{\link{NBStackedMI}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{mi}
