% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-nb-ml.R
\name{NBML}
\alias{NBML}
\title{Nonparametric Bootstrap Confidence Intervals for the Indirect Effect
using Normal Theory Maximum Likelihood
(ML nested within NB or NB(ML))}
\usage{
NBML(data, B = 5000L, mplus_bin)
}
\arguments{
\item{data}{Numeric matrix.
Output of the \code{GenData} or \code{AmputeData} functions.}

\item{B}{Positive integer.
Number of bootstrap samples.}

\item{mplus_bin}{Character string.
Path of Mplus binary.}
}
\value{
Nonparametric bootstrap confidence intervals.
\code{bc} corresponds to bias-corrected and \code{pc} corresponds to percentile
confidence intervals.
}
\description{
Nonparametric Bootstrap Confidence Intervals for the Indirect Effect
using Normal Theory Maximum Likelihood
(ML nested within NB or NB(ML))
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{MCMI}()},
\code{\link{MCML}()},
\code{\link{MINB}()},
\code{\link{NBMI}()},
\code{\link{NBStackedMI}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{ml}
