% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-nb-stacked-mi.R
\name{NBStackedMI}
\alias{NBStackedMI}
\title{Nonparametric Bootstrap Confidence Intervals for the Indirect Effect
using Multiple Imputation
(NB using Stacked MI)}
\usage{
NBStackedMI(data_mi, B = 5000L, alpha = c(0.05, 0.01, 0.001))
}
\arguments{
\item{data_mi}{List of numeric matrices.
Output of the \code{ImputeData} function
or a list of three-column data sets with imputed data.}

\item{B}{Positive integer.
Number of bootstrap samples.}

\item{alpha}{Numeric vector.
Significance level.}
}
\value{
Nonparametric bootstrap percentile confidence intervals.
}
\description{
Nonparametric Bootstrap Confidence Intervals for the Indirect Effect
using Multiple Imputation
(NB using Stacked MI)
}
\seealso{
Other Confidence Interval Functions: 
\code{\link{MCMI}()},
\code{\link{MCML}()},
\code{\link{MINB}()},
\code{\link{NBMI}()},
\code{\link{NBML}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Confidence Interval Functions}
\keyword{ci}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{mi}
