% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-sim.R
\name{Sim}
\alias{Sim}
\title{Simulation Replication}
\usage{
Sim(
  taskid,
  repid,
  output_folder,
  overwrite,
  integrity,
  params_taskid,
  alpha,
  m,
  R,
  B,
  mplus_bin,
  nbmi
)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{repid}{Positive integer.
Replication ID.}

\item{output_folder}{Character string.
Output folder.}

\item{overwrite}{Logical.
Overwrite existing output in \code{output_folder}.}

\item{integrity}{Logical.
If \code{integrity = TRUE},
check for the output file integrity when \code{overwrite = FALSE}.}

\item{params_taskid}{Data frame with a single row.
Simulation parameters for a specific \code{taskid}.}

\item{alpha}{Numeric vector.
Significance level.}

\item{m}{Positive integer.
Number of imputations.}

\item{R}{Positive integer.
Number of Monte Carlo replications.}

\item{B}{Positive integer.
Number of bootstrap samples.}

\item{mplus_bin}{Character string.
Path of Mplus binary.}

\item{nbmi}{Logical.
If \code{nbmi = TRUE}, include \code{MINB}, \code{NBMI}, and \code{StackedMINB}.}
}
\value{
The output is saved as an external file in \code{output_folder}.
}
\description{
Simulation Replication
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{manMCMedMiss}
\keyword{simulation}
