% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-sim-ampute-data.R
\name{SimAmputeData}
\alias{SimAmputeData}
\title{Simulation Replication - AmputeData}
\usage{
SimAmputeData(taskid, repid, output_folder, seed, suffix, overwrite, integrity)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{repid}{Positive integer.
Replication ID.}

\item{output_folder}{Character string.
Output folder.}

\item{seed}{Integer.
Random seed.}

\item{suffix}{Character string.
Output of \code{manMCMedMiss:::.SimSuffix()}.}

\item{overwrite}{Logical.
Overwrite existing output in \code{output_folder}.}

\item{integrity}{Logical.
If \code{integrity = TRUE},
check for the output file integrity when \code{overwrite = FALSE}.}
}
\value{
The output is saved as an external file in \code{output_folder}.
}
\description{
Simulation Replication - AmputeData
}
\details{
This function is executed via the \code{Sim} function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{ampute}
\keyword{gendata}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{simulation}
