% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-sim-fn.R
\name{SimFN}
\alias{SimFN}
\title{Simulation File Name}
\usage{
SimFN(data_type, output_type, output_folder, suffix)
}
\arguments{
\item{data_type}{Character string.
Input data type.
Valid values include
\code{"complete-00"},
\code{"mar-30"},
\code{"mar-20"},
\code{"mar-10"},
\code{"mcar-30"},
\code{"mcar-20"}, and
\code{"mcar-10"}}

\item{output_type}{Character string.
Output type.
Valid values include
\code{"data"},
\code{"fit-mi"},
\code{"fit-ml"},
\code{"mc-mi"},
\code{"mc-ml"},
\code{"mi-nb"},
\code{"nb-mi"},
\code{"nb-ml"},
\code{"sig-mi"},
\code{"sig-ml"}, and
\code{"stackedmi-nb"}}

\item{output_folder}{Character string.
Output folder.}

\item{suffix}{Character string.
Output of \code{manMCMedMiss:::.SimSuffix()}.}
}
\value{
Returns a character string file name with the \code{output_folder}
in the OS-specific format.
}
\description{
Simulation File Name
}
\keyword{manMCMedMiss}
\keyword{simulation}
