% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-sim-impute-data.R
\name{SimImputeData}
\alias{SimImputeData}
\title{Simulation Replication - ImputeData}
\usage{
SimImputeData(
  taskid,
  repid,
  output_folder,
  seed,
  suffix,
  overwrite,
  integrity,
  m,
  mplus_bin
)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{repid}{Positive integer.
Replication ID.}

\item{output_folder}{Character string.
Output folder.}

\item{seed}{Integer.
Random seed.}

\item{suffix}{Character string.
Output of \code{manMCMedMiss:::.SimSuffix()}.}

\item{overwrite}{Logical.
Overwrite existing output in \code{output_folder}.}

\item{integrity}{Logical.
If \code{integrity = TRUE},
check for the output file integrity when \code{overwrite = FALSE}.}

\item{m}{Positive integer.
Number of imputations.}

\item{mplus_bin}{Character string.
Path of Mplus binary.}
}
\value{
The output is saved as an external file in \code{output_folder}.
}
\description{
Simulation Replication - ImputeData
}
\details{
This function is executed via the \code{Sim} function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{check}
\keyword{gendata}
\keyword{impute}
\keyword{manMCMedMiss}
\keyword{simulation}
