% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-sim-joint-sig-mi.R
\name{SimJointSigMI}
\alias{SimJointSigMI}
\title{Simulation Replication - JointSigMI}
\usage{
SimJointSigMI(
  taskid,
  repid,
  output_folder,
  seed,
  suffix,
  overwrite,
  integrity,
  alpha
)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{repid}{Positive integer.
Replication ID.}

\item{output_folder}{Character string.
Output folder.}

\item{seed}{Integer.
Random seed.}

\item{suffix}{Character string.
Output of \code{manMCMedMiss:::.SimSuffix()}.}

\item{overwrite}{Logical.
Overwrite existing output in \code{output_folder}.}

\item{integrity}{Logical.
If \code{integrity = TRUE},
check for the output file integrity when \code{overwrite = FALSE}.}

\item{alpha}{Numeric vector.
Significance level.}
}
\value{
The output is saved as an external file in \code{output_folder}.
}
\description{
Simulation Replication - JointSigMI
}
\details{
This function is executed via the \code{Sim} function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{mi}
\keyword{simulation}
\keyword{test}
