% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-sum.R
\name{Sum}
\alias{Sum}
\title{Summarize Simulations}
\usage{
Sum(taskid, reps, output_folder, params_taskid, nbmi)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{reps}{Positive integer.
Number of replications.}

\item{output_folder}{Character string.
Output folder.}

\item{params_taskid}{Data frame with a single row.
Simulation parameters for a specific \code{taskid}.}

\item{nbmi}{Logical.
If \code{nbmi = TRUE}, include \code{MINB}, \code{NBMI}, and \code{StackedMINB}.}
}
\value{
The output is saved as an external file in \code{output_folder}.
}
\description{
Summarize Simulations
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{manMCMedMiss}
\keyword{summary}
