% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-sum-bind.R
\name{SumBind}
\alias{SumBind}
\title{Bind Results}
\usage{
SumBind(tasks, output_folder, data_raw_folder, params)
}
\arguments{
\item{tasks}{Positive integer.
Number of simulations tasks or cases.}

\item{output_folder}{Character string.
Output folder.}

\item{data_raw_folder}{Character string.
\code{data-raw} folder in the project directory.}

\item{params}{Data frame.
Simulation parameters for all simulation tasks or cases.}
}
\value{
Returns a dataframe of results and parameters.
}
\description{
Bind Results
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{manMCMedMiss}
\keyword{summary}
