% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-sum-hit.R
\name{SumHit}
\alias{SumHit}
\title{Extract Zero Hit and Theta Hit from Summary Output}
\usage{
SumHit(summary_output, params_taskid)
}
\arguments{
\item{summary_output}{Numeric matrix.
Output of \verb{Sum*} functions.}

\item{params_taskid}{Data frame with a single row.
Simulation parameters for a specific \code{taskid}.}
}
\value{
Returns a dataframe of results and parameters.
}
\description{
Extract Zero Hit and Theta Hit from Summary Output
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{manMCMedMiss}
\keyword{summary}
