% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-sum-joint-sig-mi.R
\name{SumJointSigMI}
\alias{SumJointSigMI}
\title{Summarize Simulations - SimJointSigMI}
\usage{
SumJointSigMI(taskid, reps, output_folder, params_taskid)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{reps}{Positive integer.
Number of replications.}

\item{output_folder}{Character string.
Output folder.}

\item{params_taskid}{Data frame with a single row.
Simulation parameters for a specific \code{taskid}.}
}
\value{
Returns a numeric matrix.
}
\description{
Summarize Simulations - SimJointSigMI
}
\details{
This function is executed via the \code{Sum} function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{mi}
\keyword{summary}
\keyword{test}
