% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-sum-mc-ml.R
\name{SumMCML}
\alias{SumMCML}
\title{Summarize Simulations - SimMCML}
\usage{
SumMCML(taskid, reps, output_folder, params_taskid)
}
\arguments{
\item{taskid}{Positive integer.
Task ID.}

\item{reps}{Positive integer.
Number of replications.}

\item{output_folder}{Character string.
Output folder.}

\item{params_taskid}{Data frame with a single row.
Simulation parameters for a specific \code{taskid}.}
}
\value{
Returns a numeric matrix.
}
\description{
Summarize Simulations - SimMCML
}
\details{
This function is executed via the \code{Sum} function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{ci}
\keyword{manMCMedMiss}
\keyword{mediation}
\keyword{ml}
\keyword{summary}
