% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-tab-describe.R
\name{TabDescribe}
\alias{TabDescribe}
\title{Descriptive Statistics of Type I Error Rate, Statistical Power, and Miss Rate}
\usage{
TabDescribe(results, type, mech, prop, n = NULL)
}
\arguments{
\item{results}{Data frame.
The package data frame \code{results()}.}

\item{type}{Character string.
Type of results to plot.
Valid values are \code{"type1"}, \code{"power"}, and \code{"miss"}
for Type I error rate, statistical power, and miss rate, respectively.}

\item{mech}{Character string.
Missing data mechanism.
Valid values are \code{"COMPLETE"}, \code{"MAR"}, and \code{"MCAR"}.}

\item{prop}{Numeric.
Proportion of missing data.
Valid values are \code{.1}, \code{.2}, \code{.3}.
Note that if \code{mech = "COMPLETE"}, \code{prop} will be set to \code{0}.}

\item{n}{Vector of positive integers.
Sample size.
Valid values are \code{50}, \code{75}, \code{100}, \code{150}, \code{200}, \code{250}, \code{500}, and \code{1000}.
If \code{n = NULL}, use all sample sizes.}
}
\value{
Returns a table of means (\code{M}),
standard deviations (\code{SD}), and
row counts (\code{N}).
}
\description{
Descriptive Statistics of Type I Error Rate, Statistical Power, and Miss Rate
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Table Functions}
\keyword{manMCMedMiss}
\keyword{table}
