% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manMCMedMiss-package.R
\docType{package}
\name{manMCMedMiss-package}
\alias{manMCMedMiss-package}
\alias{_PACKAGE}
\title{manMCMedMiss: Monte Carlo Confidence Intervals for the Indirect Effect with Missing Data}
\description{
Research compendium for the manuscript Monte Carlo Confidence Intervals for the Indirect Effect with Missing Data containing scripts used in the simulation study and other supplementary materials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/manMCMedMiss}
  \item \url{https://jeksterslab.github.io/manMCMedMiss/}
  \item Report bugs at \url{https://github.com/jeksterslab/manMCMedMiss/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID}) [contributor]
}

}
\keyword{internal}
